update constants and form helpers
