+++ ui/src/components/Runs.tsx
@@ -19,7 +19,7 @@ import Pagination from "./Pagination"
 import GenericMultiSelect from "./GenericMultiSelect"
 import RunStatusSelect from "./RunStatusSelect"
 import { FormGroup, Classes, Spinner } from "@blueprintjs/core"
-import { pageSize } from "../constants"
+import { PAGE_SIZE } from "../constants"
 import { RequestStatus } from "./Request"
 import ErrorCallout from "./ErrorCallout"
 
@@ -188,7 +188,7 @@ export const Runs: React.FunctionComponent<Props> = ({
           updatePage={updatePage}
           currentPage={currentPage}
           isLoading={isLoading}
-          pageSize={pageSize}
+          pageSize={PAGE_SIZE}
           numItems={data ? data.total : 0}
         />
       </div>
@@ -204,7 +204,10 @@ const ConnectedRuns: React.FunctionComponent<{}> = () => (
     getRequestArgs={params => ({
       params: {
         ...omit(params, "page"),
-        ...pageToOffsetLimit({ page: get(params, "page", 1), limit: pageSize }),
+        ...pageToOffsetLimit({
+          page: get(params, "page", 1),
+          limit: PAGE_SIZE,
+        }),
       },
     })}
   >
