+++ ui/src/constants.ts
@@ -1,58 +1,3 @@
-export const pageSize = 20
-
-type FieldSpec = {
-  name: string
-  label: string
-  description: string
-  initialValue: any
-}
-
-export const aliasFieldSpec: FieldSpec = {
-  name: "alias",
-  label: "alias",
-  description: "alias",
-  initialValue: "",
-}
-export const groupNameFieldSpec: FieldSpec = {
-  name: "group_name",
-  label: "Group Name",
-  description:
-    "Create a new group name or select an existing one to help searching for this task in the future.",
-  initialValue: "",
-}
-export const imageFieldSpec: FieldSpec = {
-  name: "image",
-  label: "Docker Image",
-  description: "The full URL of the Docker image and tag.",
-  initialValue: "",
-}
-export const commandFieldSpec: FieldSpec = {
-  name: "command",
-  label: "Command",
-  description: "The command for this task to execute.",
-  initialValue: "",
-}
-export const memoryFieldSpec: FieldSpec = {
-  name: "memory",
-  label: "Memory (MB)",
-  description: "The amount of memory (MB) this task needs.",
-  initialValue: 1024,
-}
-export const cpuFieldSpec: FieldSpec = {
-  name: "cpu",
-  label: "CPU (Units)",
-  description: "The amount of CPU (units) this task needs.",
-  initialValue: 512,
-}
-export const tagsFieldSpec: FieldSpec = {
-  name: "tags",
-  label: "Tags",
-  description: "",
-  initialValue: [],
-}
-export const envFieldSpec: FieldSpec = {
-  name: "env",
-  label: "Environment Variables",
-  description: "",
-  initialValue: [],
-}
+export const PAGE_SIZE = 20
+export const RUN_FETCH_INTERVAL_MS = 5000 // 5 sec
+export const LOG_FETCH_INTERVAL_MS = 10000 // 10 sec
